refactor : for avoid -Incorrect file structure or non-page files in the pages/ directory
